refactor(mlir): split definitions into separate files
